0.0.4 refactoring - : -Reorganizing files locations -create the structure of layout.jsx
