+++ src/App.jsx
@@ -1,9 +1,5 @@
-
-
-
-
 import './App.css'
-import { LoginForm } from './components/Login'
+import { LoginForm } from './Pages/Login'
 import './index.css'
 
 
@@ -13,9 +9,6 @@ function App() {
 
   return (
     <>
-    {/* <h1 className="text-3xl font-bold underline">
-     
-    </h1> */}
     <LoginForm/>
     </>
   )
